Update: Project classes structure

+++ app/src/main/java/com/joshdev/smartpocket/ui/activities/currency/CurrencyViewModel.kt
@@ -69,7 +69,9 @@ class CurrencyViewModel : ViewModel() {
     fun deleteCurrency() {
         selectedCurrency.value?.let { currency ->
             viewModelScope.launch(Dispatchers.IO) {
-                operations.deleteItem<Currency, CurrencyRealm>(currency.id)
+                if (currency.name != "USD") {
+                    operations.deleteItem<Currency, CurrencyRealm>(currency.id)
+                }
             }
         }
     }
